now passing test_tpch_workload

+++ benchmark/tests/integtest_benchmark.py
@@ -2,11 +2,20 @@ import shutil
 import unittest
 from pathlib import Path
 
-from gymlib.symlinks_paths import get_tables_symlink_path
+from gymlib.symlinks_paths import (
+    get_tables_symlink_path,
+    get_workload_suffix,
+    get_workload_symlink_path,
+)
 
 # It's ok to import private functions from the benchmark module because this is an integration test.
-from benchmark.tpch.cli import _tpch_tables
-from util.workspace import DBGymWorkspace, get_workspace_path_from_config
+from benchmark.tpch.cli import _tpch_tables, _tpch_workload
+from benchmark.tpch.constants import DEFAULT_TPCH_SEED
+from util.workspace import (
+    DBGymWorkspace,
+    fully_resolve_path,
+    get_workspace_path_from_config,
+)
 
 
 class TestBenchmark(unittest.TestCase):
@@ -15,11 +24,13 @@ class TestBenchmark(unittest.TestCase):
     def setUp(self) -> None:
         workspace_path = get_workspace_path_from_config(TestBenchmark.DBGYM_CONFIG_PATH)
         # Get a clean start each time.
-        shutil.rmtree(workspace_path)
+        if workspace_path.exists():
+            shutil.rmtree(workspace_path)
         self.workspace = DBGymWorkspace(workspace_path)
 
-    # def tearDown(self) -> None:
-    #     shutil.rmtree(self.workspace.dbgym_workspace_path)
+    def tearDown(self) -> None:
+        if self.workspace.dbgym_workspace_path.exists():
+            shutil.rmtree(self.workspace.dbgym_workspace_path)
 
     def test_tpch_tables(self) -> None:
         scale_factor = 0.01
@@ -29,6 +40,27 @@ class TestBenchmark(unittest.TestCase):
         self.assertFalse(tables_path.exists())
         _tpch_tables(self.workspace, scale_factor)
         self.assertTrue(tables_path.exists())
+        self.assertTrue(fully_resolve_path(tables_path).exists())
+
+    def test_tpch_workload(self) -> None:
+        scale_factor = 0.01
+        workload_path = get_workload_symlink_path(
+            self.workspace.dbgym_workspace_path,
+            "tpch",
+            scale_factor,
+            get_workload_suffix(
+                "tpch",
+                seed_start=DEFAULT_TPCH_SEED,
+                seed_end=DEFAULT_TPCH_SEED,
+                query_subset="all",
+            ),
+        )
+        self.assertFalse(workload_path.exists())
+        _tpch_workload(
+            self.workspace, DEFAULT_TPCH_SEED, DEFAULT_TPCH_SEED, "all", scale_factor
+        )
+        self.assertTrue(workload_path.exists())
+        self.assertTrue(fully_resolve_path(workload_path).exists())
 
 
 if __name__ == "__main__":
